refactor: organizes Tab exports in their respective folder
