+++ src/components/ui/tab/index.ts
@@ -1,10 +1,3 @@
-import type { ComponentOptions } from 'vue';
-import Tab from './Tab.vue';
-import TabItem from './TabItem.vue';
-
-export default {
-  install(app: ComponentOptions, { prefix }: { prefix: string }) {
-    app.component(`${prefix}Tab`, Tab);
-    app.component(`${prefix}TabItem`, TabItem);
-  },
-};
+export { default as Tab } from './Tab.vue';
+export { default as TabItem } from './TabItem.vue';
+export type { TabProps, TabItemProps, TabProviderInterface } from './types';
