+++ src/components/ui/tab/types.ts
@@ -3,3 +3,13 @@ export interface TabProviderInterface {
   activeTabIndex: number | string;
   active(index: number | string, evt?: unknown): void;
 }
+
+export interface TabProps {
+  modelValue: any;
+}
+
+export interface TabItemProps {
+  label?: string;
+  index?: any;
+  badge?: string | number;
+}
